Add support for loading spinner in alert boxes
The pattern of placing a spinner in an alert (styled info) to notify
that something is loading is very common for me. I realize the material
design suggests to use a progress bar but hack.css doesn't yet support
an indefinite progress bar. Till then I will often use an alert box
with a spinner.

The former style had no display type so it defaulted to block for
`<div>`s and inline for `<span>`s. The problem with a div is it means
the text next to a spinner drop to another row. This caused rendering
the spinner in an alert box looking off. It also forced the user to
explicitly set the display to inline-block or float.

For me that last think I conclude when using a CSS library like this is
to manually understand floating a spinner or changing it's display type.
Normally I would simply define the spinner as a `<span>` not a `<div>`.
Since inline elements have no width or height it renders awful.

To make things easier this adds some opinions to the display of a span
or div with .loading class on it. However, to prevent changes to style
when upgrading the spinner in a button defined as a div will stay float
left.

The example docs were updated and I moved the spinner to the right in
the button example because in almost all cases where I've seen loading
buttons the spinner has been on the right.